+++ crates/mail-uniffi/src/mail/avatar_information.rs
@@ -1,5 +1,6 @@
-use crate::mail::datatypes::{AvatarInformation, MessageAddress};
-use proton_mail_common::avatar::AvatarInformation as RealAvatarInformation;
+use crate::core::datatypes::AvatarInformation;
+use crate::mail::datatypes::MessageAddress;
+use proton_core_common::datatypes::AvatarInformation as RealAvatarInformation;
 use proton_mail_common::datatypes::MessageAddress as RealMessageAddress;
 
 /// Creates an [`AvatarInformation`] by taking then display name and email address
@@ -9,7 +10,9 @@ pub fn avatar_information_from_name_and_email(
     display_name: &str,
     email: &str,
 ) -> AvatarInformation {
-    RealAvatarInformation::build(display_name, email).into()
+    RealAvatarInformation::from(display_name)
+        .or_else(email)
+        .into()
 }
 
 /// Creates an [`AvatarInformation`] struct using the details of the first [`MessageAddress`] in the provided slice.
@@ -21,11 +24,11 @@ pub fn avatar_information_from_message_addresses(
         .iter()
         .map(|addr| RealMessageAddress::from(addr.clone()))
         .collect();
-    RealAvatarInformation::from_message_addresses(&addresses).into()
+    RealMessageAddress::avatar_info(&addresses).into()
 }
 
 /// Creates an [`AvatarInformation`] struct using a [`MessageAddress`].
 #[uniffi::export]
 pub fn avatar_information_from_message_address(address: &MessageAddress) -> AvatarInformation {
-    RealAvatarInformation::from_message_address(&address.clone().into()).into()
+    RealAvatarInformation::from(RealMessageAddress::from(address.clone())).into()
 }