+++ crates/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+### Added
+  - `contact_list` method which returns new set of data objects representing contact list
+
 ## [0.13.0] - 2024-10-11
 
 ### Added
@@ -27,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Blocking forms of `get_account[s]` and `get_session[s]`
   - Blocking forms of `get_account_state` and `get_session_state`
   - Blocking form of `get_primary_account`
-  - `LabelAs` action set labels to a group of messages 
+  - `LabelAs` action set labels to a group of messages
 
 ## [0.11.56] - 2024-10-03